moved components to own directory
