+++ src/components/participants/participant/avatar.tsx
