+++ src/components/participants/participant/icons/menu-items/lock.tsx
