+++ src/components/participants/participant/icons/menu-items/lowerHand.tsx
