+++ src/components/participants/participant/icons/menu-items/removeUser.tsx
