+++ src/components/participants/participant/icons/menu-items/switchPresenter.tsx
