+++ src/components/participants/participant/icons/menu-items/webcam.tsx
