+++ src/components/participants/participant/icons/mic.tsx
