+++ src/components/participants/participant/icons/presenterIcon.tsx
