+++ src/components/participants/participant/icons/raiseHand.tsx
