+++ src/components/participants/participant/icons/visibility.tsx
