+++ src/components/participants/participant/icons/webcam.tsx
