+++ src/components/participants/participant/name.tsx
